Refactor the facebook scraper into its own named files so that there is space for other scrapers
