+++ scripts/src/models/primitive.ts
@@ -1,0 +1,7 @@
+export type UUID = string;
+export type FacebookID = string;
+export type TimeStamp = string;
+export type Url = string;
+export type WholeNumber = number;
+export type PhoneNumber = string;
+export type Email = string;
