+++ scripts/src/runners/index.ts
@@ -1,0 +1,1 @@
+export * from "./facebookRunner";
